Linux: Use ptrdiff_t for __rseq_offset
This matches the data size initial-exec relocations use on most
targets.

+++ sysdeps/nptl/dl-tls_init_tp.c
@@ -34,7 +34,7 @@ rtld_hidden_data_def (__nptl_set_robust_list_avail)
 
 const unsigned int __rseq_flags;
 const unsigned int __rseq_size attribute_relro;
-const int __rseq_offset attribute_relro;
+const ptrdiff_t __rseq_offset attribute_relro;
 
 void
 __tls_init_tp (void)
@@ -89,7 +89,7 @@ __tls_init_tp (void)
        all targets support __thread_pointer, so set __rseq_offset only
        if thre rseq registration may have happened because RSEQ_SIG is
        defined.  */
-    extern int offset __asm__ ("__rseq_offset");
+    extern ptrdiff_t offset __asm__ ("__rseq_offset");
     offset = (char *) &pd->rseq_area - (char *) __thread_pointer ();
 #endif
   }
